Added createHMAC to TlsCrypto Added use of the same.

+++ tls/src/main/java/org/bouncycastle/tls/crypto/bc/BcTlsCrypto.java
@@ -502,7 +502,7 @@ public class BcTlsCrypto
         }
     }
 
-    public TlsHMAC createHMAC(short hashAlgorithm)
+    public TlsHMAC createHMAC(int hashAlgorithm)
     {
         switch (hashAlgorithm)
         {
